+++ .idea/inspectionProfiles/profiles_settings.xml
@@ -1,0 +1,5 @@
+<component name="InspectionProjectProfileManager">
+  <settings>
+    <option name="PROJECT_PROFILE" />
+  </settings>
+</component>
