--- NBA/processing.py
@@ -1,2 +1,0 @@
-def do_calculation(a, b):
-    return a + b
