--- NBA/templates/add.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Displays page for adding two numbers.</title>
-</head>
-<body>
-    {{errors}}
-    <p>Enter your numbers:</p>
-    <form method="post" action="http://127.0.0.1:5000/add">
-        <p><input name="number1" /></p>
-		<p><input name="number2" /></p>
-		<p><input type="submit" value="Do calculation" /></p>
-    </form>
-</body>
-</html>
