--- NBA/templates/homepage.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Homepage for NBA daily stats site.</title>
-</head>
-<body>
-    <h1>{{ date }} Matchups!</h1>
-</body>
-</html>
