+++ NBA/web_app/templates/results.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Results from adding page.</title>
+</head>
+<body>
+    <p>The result is: {{ result }}</p>
+    <a href="/add">Try again?</a>
+</body>
+</html>
